+++ tools/parse_input_validator.py
@@ -1,6 +1,7 @@
 def parse_input_validator(func):
     def inner(*args, **kwargs):
-        if(len(args[0]) == 0):
-            return (None, None)
+        if len(args[0]) == 0:
+            return None, None
         return func(*args, **kwargs)
+
     return inner
